refactor!: because I just wanted ts...
