--- package/src/options/options.html
@@ -1,77 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-	<meta charset="utf-8">
-	<link rel="stylesheet" href="style.css" />
-</head>
-
-<body>
-	<form>
-		<div>
-			<label>
-				<input type="checkbox" id="switchToTabAfterMoving" data-option="true">
-				Switch to active tab after move to another window<br>
-				<i>(middle clicking on the toolbar icon ignores this options)</i>
-			</label>
-		</div>
-
-		<div>
-			<label>
-				<input type="checkbox" id="showLastWindowIDBadge" data-option="true">
-				Show previously active window ID in the badge
-			</label>
-		</div>
-
-		<div>
-			<label>
-				Tab travel distance:
-				<input type="number" id="tabTravelDistance" data-option="true" min="0" max="10">
-			</label>
-		</div>
-
-		<div>
-			<label>
-				Recent tab timeout (seconds):
-				<input type="number" id="recentTabTimeout" data-option="true" min="0">
-			</label>
-		</div>
-
-		<div class="text-input">
-			<hr>
-			Always move below containers to a new window:
-			<br>
-
-			<table id="identityTable">
-				<tr>
-					<td>Set:</td>
-					<td id="setContainers">
-						<input type="hidden" id="moveableContainers" data-option="true">
-					</td>
-				</tr>
-
-				<tr>
-					<td>Available:</td>
-					<td id="availableContainers"></td>
-				</tr>
-			</table>
-			<hr>
-		</div>
-
-		<div>
-			<label>
-				<input type="checkbox" id="debugMode" data-option="true">
-				debug
-			</label>
-		</div>
-
-		<div class="form-buttons">
-			<button id="submit" type="submit">Save</button>
-			<button id="reset" type="button">Reset</button>
-		</div>
-	</form>
-
-	<script src="options.js"></script>
-</body>
-
-</html>
