+++ src/background/src/utils/Tab.ts
@@ -1,0 +1,9 @@
+export class TabUtils {
+
+	async getCurrentTabInWindow(): Promise<browser.tabs.Tab[]> {
+		return await browser.tabs.query({
+			active: true,
+			windowId: browser.windows.WINDOW_ID_CURRENT
+		});
+	}
+}
