+++ src/popup/popup.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html xmlns="http://www.w3.org/1999/xhtml">
+
+<head>
+	<meta charset="utf-8">
+	<style type="text/css"></style>
+	<link rel="stylesheet" href="popup.css" />
+</head>
+
+<body>
+	<script src="popup.js"></script>
+</body>
+
+</html>
